build: run go mod tidy wallet
