+++ services/go.mod
@@ -27,6 +27,7 @@ require (
 	github.com/jmoiron/sqlx v1.3.5
 	github.com/lib/pq v1.10.9
 	github.com/linkedin/goavro v2.1.0+incompatible
+	github.com/pkg/errors v0.9.1
 	github.com/prometheus/client_golang v1.13.0
 	github.com/redis/go-redis/v9 v9.7.0
 	github.com/rs/zerolog v1.28.0
