+++ services/wallet/controllers_v4_test.go
@@ -77,11 +77,11 @@ func (suite *WalletControllersV4TestSuite) TestCreateBraveWalletV4_Success() {
 		Validate(gomock.Any(), geoCountry).
 		Return(true, nil)
 
-	service, err := wallet.InitService(storage, nil, nil, nil, reputationClient, nil, locationValidator, backoff.Retry, nil, nil, wallet.DAppConfig{})
+	service, err := wallet.InitService(storage, nil, nil, nil, nil, reputationClient, nil, locationValidator, backoff.Retry, nil, nil, wallet.DAppConfig{})
 	suite.Require().NoError(err)
 
 	router := chi.NewRouter()
-	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw())
+	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw(), noOpMw())
 
 	data := wallet.V4Request{
 		GeoCountry: geoCountry,
@@ -124,11 +124,11 @@ func (suite *WalletControllersV4TestSuite) TestCreateBraveWalletV4_GeoCountryDis
 		Validate(gomock.Any(), gomock.Any()).
 		Return(false, nil)
 
-	service, err := wallet.InitService(nil, nil, nil, nil, nil, nil, locationValidator, backoff.Retry, nil, nil, wallet.DAppConfig{})
+	service, err := wallet.InitService(nil, nil, nil, nil, nil, nil, nil, locationValidator, backoff.Retry, nil, nil, wallet.DAppConfig{})
 	suite.Require().NoError(err)
 
 	router := chi.NewRouter()
-	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw())
+	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw(), noOpMw())
 
 	data := wallet.V4Request{
 		GeoCountry: "AF",
@@ -176,11 +176,11 @@ func (suite *WalletControllersV4TestSuite) TestCreateBraveWalletV4_WalletAlready
 		Validate(gomock.Any(), geoCountry).
 		Return(true, nil)
 
-	service, err := wallet.InitService(storage, nil, nil, nil, nil, nil, locationValidator, nil, nil, nil, wallet.DAppConfig{})
+	service, err := wallet.InitService(storage, nil, nil, nil, nil, nil, nil, locationValidator, nil, nil, nil, wallet.DAppConfig{})
 	suite.Require().NoError(err)
 
 	router := chi.NewRouter()
-	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw())
+	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw(), noOpMw())
 
 	data := wallet.V4Request{
 		GeoCountry: geoCountry,
@@ -247,11 +247,11 @@ func (suite *WalletControllersV4TestSuite) TestCreateBraveWalletV4_ReputationCal
 		Validate(gomock.Any(), gomock.Any()).
 		Return(true, nil)
 
-	service, err := wallet.InitService(storage, nil, nil, nil, reputationClient, nil, locationValidator, backoff.Retry, nil, nil, wallet.DAppConfig{})
+	service, err := wallet.InitService(storage, nil, nil, nil, nil, reputationClient, nil, locationValidator, backoff.Retry, nil, nil, wallet.DAppConfig{})
 	suite.Require().NoError(err)
 
 	router := chi.NewRouter()
-	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw())
+	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw(), noOpMw())
 
 	data := wallet.V4Request{
 		GeoCountry: "AF",
@@ -297,7 +297,7 @@ func (suite *WalletControllersV4TestSuite) TestUpdateBraveWalletV4_Success() {
 		UpsertReputationSummary(gomock.Any(), gomock.Any(), gomock.Any()).
 		Return(nil)
 
-	service, err := wallet.InitService(storage, nil, nil, nil, reputationClient, nil, nil, backoff.Retry, nil, nil, wallet.DAppConfig{})
+	service, err := wallet.InitService(storage, nil, nil, nil, nil, reputationClient, nil, nil, backoff.Retry, nil, nil, wallet.DAppConfig{})
 	suite.Require().NoError(err)
 
 	// create rewards wallet with public key
@@ -318,7 +318,7 @@ func (suite *WalletControllersV4TestSuite) TestUpdateBraveWalletV4_Success() {
 	suite.Require().NoError(err)
 
 	router := chi.NewRouter()
-	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw())
+	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw(), noOpMw())
 
 	data := wallet.V4Request{
 		GeoCountry: "AF",
@@ -347,7 +347,7 @@ func (suite *WalletControllersV4TestSuite) TestUpdateBraveWalletV4_VerificationM
 	storage, err := wallet.NewWritablePostgres("", false, "")
 	suite.NoError(err)
 
-	service, err := wallet.InitService(storage, nil, nil, nil, nil, nil, nil, backoff.Retry, nil, nil, wallet.DAppConfig{})
+	service, err := wallet.InitService(storage, nil, nil, nil, nil, nil, nil, nil, backoff.Retry, nil, nil, wallet.DAppConfig{})
 	suite.Require().NoError(err)
 
 	publicKey, privateKey, err := httpsignature.GenerateEd25519Key(nil)
@@ -356,7 +356,7 @@ func (suite *WalletControllersV4TestSuite) TestUpdateBraveWalletV4_VerificationM
 	paymentID := uuid.NewV5(wallet.ClaimNamespace, publicKey.String()).String()
 
 	router := chi.NewRouter()
-	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw())
+	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw(), noOpMw())
 
 	data := wallet.V4Request{
 		GeoCountry: "AF",
@@ -385,7 +385,7 @@ func (suite *WalletControllersV4TestSuite) TestUpdateBraveWalletV4_PaymentIDMism
 	storage, err := wallet.NewWritablePostgres("", false, "")
 	suite.NoError(err)
 
-	service, err := wallet.InitService(storage, nil, nil, nil, nil, nil, nil, backoff.Retry, nil, nil, wallet.DAppConfig{})
+	service, err := wallet.InitService(storage, nil, nil, nil, nil, nil, nil, nil, backoff.Retry, nil, nil, wallet.DAppConfig{})
 	suite.Require().NoError(err)
 
 	// create rewards wallet with public key
@@ -406,7 +406,7 @@ func (suite *WalletControllersV4TestSuite) TestUpdateBraveWalletV4_PaymentIDMism
 	suite.Require().NoError(err)
 
 	router := chi.NewRouter()
-	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw())
+	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw(), noOpMw())
 
 	data := wallet.V4Request{
 		GeoCountry: "AF",
@@ -452,7 +452,7 @@ func (suite *WalletControllersV4TestSuite) TestUpdateBraveWalletV4_GeoCountryAlr
 		UpsertReputationSummary(gomock.Any(), gomock.Any(), gomock.Any()).
 		Return(errorBundle)
 
-	service, err := wallet.InitService(storage, nil, nil, nil, reputationClient, nil, nil, backoff.Retry, nil, nil, wallet.DAppConfig{})
+	service, err := wallet.InitService(storage, nil, nil, nil, nil, reputationClient, nil, nil, backoff.Retry, nil, nil, wallet.DAppConfig{})
 	suite.Require().NoError(err)
 
 	// create rewards wallet with public key
@@ -473,7 +473,7 @@ func (suite *WalletControllersV4TestSuite) TestUpdateBraveWalletV4_GeoCountryAlr
 	suite.Require().NoError(err)
 
 	router := chi.NewRouter()
-	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw())
+	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw(), noOpMw())
 
 	data := wallet.V4Request{
 		GeoCountry: "AF",
@@ -517,7 +517,7 @@ func (suite *WalletControllersV4TestSuite) TestUpdateBraveWalletV4_ReputationCal
 		UpsertReputationSummary(gomock.Any(), gomock.Any(), gomock.Any()).
 		Return(errReputation)
 
-	service, err := wallet.InitService(storage, nil, nil, nil, reputationClient, nil, nil, backoff.Retry, nil, nil, wallet.DAppConfig{})
+	service, err := wallet.InitService(storage, nil, nil, nil, nil, reputationClient, nil, nil, backoff.Retry, nil, nil, wallet.DAppConfig{})
 	suite.Require().NoError(err)
 
 	// create rewards wallet with public key
@@ -538,7 +538,7 @@ func (suite *WalletControllersV4TestSuite) TestUpdateBraveWalletV4_ReputationCal
 	suite.Require().NoError(err)
 
 	router := chi.NewRouter()
-	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw())
+	wallet.RegisterRoutes(ctx, service, router, noOpHandler(), noOpMw(), noOpMw())
 
 	data := wallet.V4Request{
 		GeoCountry: "AF",
@@ -582,7 +582,7 @@ func (suite *WalletControllersTestSuite) TestGetWalletV4() {
 
 	allowList := storage.NewAllowList()
 
-	service, _ := wallet.InitService(pg, nil, nil, allowList, nil, nil, nil, nil, nil, nil, wallet.DAppConfig{})
+	service, _ := wallet.InitService(pg, nil, nil, allowList, nil, nil, nil, nil, nil, nil, nil, wallet.DAppConfig{})
 
 	handler := handlers.AppHandler(wallet.GetWalletV4(service))
 
@@ -625,7 +625,7 @@ func (suite *WalletControllersTestSuite) TestGetWalletV4_Not_Whitelisted() {
 
 	allowList := storage.NewAllowList()
 
-	service, _ := wallet.InitService(pg, nil, nil, allowList, nil, nil, nil, nil, nil, nil, wallet.DAppConfig{})
+	service, _ := wallet.InitService(pg, nil, nil, allowList, nil, nil, nil, nil, nil, nil, nil, wallet.DAppConfig{})
 
 	handler := handlers.AppHandler(wallet.GetWalletV4(service))
